refactor: decouple flight_event and web_tz_switcher modules
