+++ flight_event/__manifest__.py
@@ -14,10 +14,11 @@
         "base",
         "flight",
         "web",
-        "web_tz_switcher",
     ],
     "assets": {
         "web.assets_backend": [
+            # Shared hooks
+            "flight_event/static/src/hooks/*.js",
             # Shared time input cell (used by both matrix and summary widgets)
             "flight_event/static/src/components/flight_time_input_cell/flight_time_input_cell.scss",
             "flight_event/static/src/components/flight_time_input_cell/flight_time_input_cell.js",
